Move statistics to elastic-job-lite-console module
