+++ elastic-job-lite-console/src/main/java/org/apache/shardingsphere/elasticjob/lite/console/dao/statistics/type/job/JobExecutionTypeStatistics.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,21 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.tracing.rdb.statistics.type.job;
+package org.apache.shardingsphere.elasticjob.lite.console.dao.statistics.type.job;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 /**
- * Job type statistics.
+ * Job execution type statistics.
  */
 @Getter
 @AllArgsConstructor
-public final class JobTypeStatistics {
+public final class JobExecutionTypeStatistics {
     
-    private int scriptJobCount;
+    private int transientJobCount;
     
-    private int simpleJobCount;
-    
-    private int dataflowJobCount;
+    private int daemonJobCount;
 }
