+++ elastic-job-lite-console/src/main/java/org/apache/shardingsphere/elasticjob/lite/console/dao/statistics/type/task/TaskRunningStatistics.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.tracing.rdb.statistics.type.job;
+package org.apache.shardingsphere.elasticjob.lite.console.dao.statistics.type.task;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -24,16 +24,16 @@ import lombok.RequiredArgsConstructor;
 import java.util.Date;
 
 /**
- * Job register statistics.
+ * Task running statistics.
  */
 @Getter
 @RequiredArgsConstructor
 @AllArgsConstructor
-public final class JobRegisterStatistics {
+public final class TaskRunningStatistics {
     
     private long id;
     
-    private final int registeredCount;
+    private final int runningCount;
     
     private final Date statisticsTime;
